requirements.txt
langchain
openai
faiss-cpu
streamlit
streamlit-chat
tiktoken
weaviate-client
black
isort
Flask
transformers
gradio
pypdf
unstructured
pdfminer.six
pandoc
pypandoc
fastapi
black
isort
websockets
pydantic
uvicorn
jinja2
faiss-cpu
bs4
unstructured
libmagic